move file managment to FIleManager, create chunk size and a overrideable prop, attach executeor to TUSClient calls, add Cacnel to TUSEexecutor
